Fix type of intensity_tolerance variable of structure.
git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1941 3c61e6fb-ae43-4039-8cf7-c895b3257837

+++ include/sbia/basis/testdriver.h
@@ -121,7 +121,7 @@ MultiStringArg add_before_env_with_sep(
 struct RegressionTest {
     string       test_image_file;
     string       baseline_image_file;
-    float        intensity_tolerance;
+    double       intensity_tolerance;
     unsigned int max_number_of_differences;
     unsigned int tolerance_radius;
 };
